requirements.txt
fastapi
uvicorn
easyocr
pillow
numpy
